rename RobotWrapper trait to Robot
