+++ examples/generate_cartesian_velocity_motion.rs
@@ -6,9 +6,7 @@ use std::time::Duration;
 
 use clap::Parser;
 
-use franka::{
-    CartesianVelocities, Fr3, FrankaResult, MotionFinished, Panda, RobotState, RobotWrapper,
-};
+use franka::{CartesianVelocities, Fr3, FrankaResult, MotionFinished, Panda, Robot, RobotState};
 
 /// An example showing how to generate a Cartesian velocity motion.
 ///
@@ -37,7 +35,7 @@ fn main() -> FrankaResult<()> {
     }
 }
 
-fn generate_motion<R: RobotWrapper>(mut robot: R) -> FrankaResult<()> {
+fn generate_motion<R: Robot>(mut robot: R) -> FrankaResult<()> {
     robot.set_default_behavior()?;
     println!("WARNING: This example will move the robot! Please make sure to have the user stop button at hand!");
     println!("Press Enter to continue...");
