+++ examples/generate_consecutive_motions.rs
@@ -7,7 +7,7 @@ use std::time::Duration;
 use clap::Parser;
 
 use franka::exception::FrankaException;
-use franka::{Fr3, FrankaResult, JointVelocities, MotionFinished, Panda, RobotState, RobotWrapper};
+use franka::{Fr3, FrankaResult, JointVelocities, MotionFinished, Panda, Robot, RobotState};
 
 /// An example showing how to execute consecutive motions with error recovery.
 ///
@@ -37,7 +37,7 @@ fn main() -> FrankaResult<()> {
     }
 }
 
-fn generate_motion<R: RobotWrapper>(mut robot: R) -> FrankaResult<()> {
+fn generate_motion<R: Robot>(mut robot: R) -> FrankaResult<()> {
     robot.set_default_behavior()?;
     println!("WARNING: This example will move the robot! Please make sure to have the user stop button at hand!");
     println!("Press Enter to continue...");
