+++ examples/generate_joint_velocity_motion.rs
@@ -6,7 +6,7 @@ use std::time::Duration;
 
 use clap::Parser;
 
-use franka::{Fr3, FrankaResult, JointVelocities, MotionFinished, Panda, RobotState, RobotWrapper};
+use franka::{Fr3, FrankaResult, JointVelocities, MotionFinished, Panda, Robot, RobotState};
 
 /// An example showing how to generate a joint velocity motion.
 ///
@@ -35,7 +35,7 @@ fn main() -> FrankaResult<()> {
     }
 }
 
-fn generate_motion<R: RobotWrapper>(mut robot: R) -> FrankaResult<()> {
+fn generate_motion<R: Robot>(mut robot: R) -> FrankaResult<()> {
     robot.set_default_behavior()?;
     println!("WARNING: This example will move the robot! Please make sure to have the user stop button at hand!");
     println!("Press Enter to continue...");
