+++ examples/print_joint_poses.rs
@@ -3,7 +3,7 @@
 use clap::Parser;
 use nalgebra::Matrix4;
 
-use franka::{Fr3, Frame, FrankaResult, Panda, RobotModel, RobotWrapper};
+use franka::{Fr3, Frame, FrankaResult, Panda, Robot, RobotModel};
 
 /// An example showing how to use the model library that prints the transformation
 /// matrix of each joint with respect to the base frame.
@@ -31,7 +31,7 @@ fn main() -> FrankaResult<()> {
     }
 }
 
-fn generate_motion<R: RobotWrapper>(mut robot: R) -> FrankaResult<()> {
+fn generate_motion<R: Robot>(mut robot: R) -> FrankaResult<()> {
     let model = robot.load_model(false)?;
     let state = robot.read_once()?;
     let frames = vec![
